Rename `closed_at` column on Projects to `completed_at`
We are going to change the wording of "closed" projects on the front end to
"completed". For clarity, we have decided to keep the front and back end
terminology the same, which means renaming the `closed_at` column on the
Project model to `completed_at`.

This change also means changes to the relevant controllers, methods, tests,
views and text strings.