Set up accessibility tool
`axe-core-rspec` has been used as the accessibility tool for our automated
accessibility tests.
I have made sure that the test currently does not run within the default
test suite, as this is something that we want to be able to turn on when
we want to test out accessibility needs and to also prevent the blocking
of work being merged in when these accessibility tests fail.