Split Project index page into "In progress" (open) and "Completed" tabs
We want to make completed projects more easily findable on the index page.

Previously, completed (or "closed") projects were pushed to the end of the
project list, so it wasn't entirely obvious where they were. This commit splits
the index page into two actions, and presents them on separate tabs on the
project index page.

The first tab is "In progress" projects, i.e projects which are open (do not
have a `completed_at` date). The second tab is for "Completed" projects. Both
tabs have their own route & action.

As a side effect of this, the layout for the index page has been put into a
partial, to be included in the views for both the `index` and `completed`
actions.

The relevant tests have been updated, and a new test for the completed project
view has been added.