Rename `target_completion_date` to `provisional_conversion_date`
We have renamed the `target_completion_date` column on Projects to
`provisional_conversion_date`. The content will be changing on the front end
and it makes sense to keep the front and back end terms in sync.

This commit includes the migration, model validations and i18n keys associated
with validations.