Allow ci to continue on error
Within our continuous integration tests, the accessibility checks are
not required to pass for work to be merged in.
The continue-on-error action prevents a job from failing when a step
fails.
We currently expect the accessibility checks to fail right now.
This is why we are taking steps to ensure the red check mark doesn't
show for every PR to prevent any merging blocks.