Update Sentry configuration
We have decided not to send Sentry data in our local development
environments. Sentry is billed per 'event' so we want to minimise how
many we send.

Here we remove the two environment variables relating to Sentry for
local development and guard against configuring sentry unless we are in
a production environment i.e one of our live environments.

We have to update the environment banner helper to provide a default
when the SENTRY_ENV is not present.