Rename Project to ConversionProject
We are going to divide the "projects" on this application into Conversion
Projects and Transfers/Significant Change Projects. The Conversion Projects will
be divided into Voluntary and Involuntary.

As a first step, we have decided to rename the Project model to
ConversionProject.

This was the smallest wecould get this commit while still keeping all the tests
green! We have:

- Renamed the Project model to ConversionProject
- Aliased the `conversion_project` attribute on Notes, Tasks and Sections as
`project`, in order the minimise the amount of edits needed to other files
- Renamed some of the essential i18n keys from `project` to `conversion_project`
- Edited the routes associated with ConversionProjects
- Renamed the Project factory to ConversionProject