Add conversion project
We know that there are three types of project that the service must
support:

- converions
- transfers
- changes

Up until now we have made no distinction of the type of project, as we
would like to introduce two kinds of converison projects, now is the
time to do so.

We plan to use Single Table Inheritance to achieve this. Here we add the
type column to projects that allows it to become a conversion project.