Add conversion details
Now that we have a conversion project, we want to be able to store
details about them but not use the projects table to do so, the projects
table only contains shared attributes.

Here we create a new converion_projects table to store attrbutes and set
up an associations between the two.

We also scaffold the two types of converison projects:

- voluntary (with a approval order)
- involuntary (with a directive approval order)

We don't use them yet, we just like having them around.