Add a way to create involuntary conversion projects
We want to be able to create Involuntary conversion projects. In order to do
this, we have created a Conversion Projects Controller, and an Involuntary
Conversion Projects Controller which inherits from it.

When the involuntary conversion project create form is posted to the
Involuntary Conversion Projects controller, it creates a Conversion Project,
with an Involuntary Conversion Details object.

This commit also includes a small refactor to the project i18n keys, which allows
them to be used for both projects and conversion projects.