Fix the back link accessibility violations
Previously the back link had no landmark which caused axe to throw an
error. By wrapping the link in a `<nav>` element with an `aria-label` we
provide the missing landmark and fix the problem.
We created a partial file to encapsulate this work.

https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Roles/navigation_role

The govuk design system is missing this information.