Use `bulk` API endpoints for Trust and Establishment
Currently, our pages load very slowly as we send an API request for each Trust
and Establishment related to a project. To help with this, we have added `bulk`
endpoints to the API so that we can retrieve multiple Trusts and Establishments
in a single request.

These new endpoints are significantly faster than the single resource
endpoints. For now, this swaps out the API requests, but keeps them separate. We
will see if this has a big enough impact on page load before doing the work to
fetch all of the Trusts and Establishments in one go.