Allow Simplecov to not run
During our accessibility checks we don't want to run simplecov as the
tests will not result in 100% coverage.

Here we add the ability to set 'NO_COVERAGE=true' to stop simplecov
from running.

Setting false or not setting the variable results in simplecov running
which is the expexted behaviour and means we don't have to change
anything else.

With our accessibility test now passing we can disable
`continue-on-error` for the GitHub Action now.