Only test with Conversion projects in the test suite
From now on, all projects we create in the app will be Conversion projects.
Until we add more types of projects, conversion projects will be the default.
We do not want to run tests with "plain" projects as they aren't used directly
by any of the application functions.

Change the project factory to only create Conversion::Projects.

The change in `spec/models/project_spec.rb:284` is necessary because when we
call `Project#index` the projects returned will be Project objects, while the
projects created by the project factory will be Conversion::Projects. Due to
STI these are the same type of objects, but with different class names. We have
amended this test to check the returned objects have the same attributes.

Additionally, move the Project factory into a `conversion` folder