Add method to get multiple establishments by URNs to client
We want to be able to fetch multiple establishments in a single request on the
projects index page.

This adds a `get_establishments` method to the API client, which calls the new
`bulk` endpoint on the API.

It also refactors the `get_establishment` method to use the `bulk` endpoint to
keep things DRY, with the benefit that performance is better on the bulk
endpoint than on the single resource endpoint

The Faraday params encoder is set as the `Faraday::FlatParamsEncoder`. This
enables us to to send the same URL parameter multiple times with different
values, and is needed because the API `bulk` endpoints have repeating query
strings (e.g., `?urn=123456&urn234567&urn=345678`).