Add method to get multiple trusts by UKPRNs to client
We want to be able to fetch multiple trusts in a single request on the projects
index page.

This adds a `get_trusts` method to the API client, which calls the new `bulk`
endpoint on the API.

It also refactors the `get_trust` method to use the `bulk` endpoint to keep
things DRY, with the benefit that performance is better on the bulk endpoint
than on the single resource endpoint.