Add EstablishmentsFetcher service
We need to be able to map the establishments fetched from the
`get_establishments` method of the client to the projects.

This adds a EstablishmentsFetcher service to perform this. The URN must be
added to the API response so that we are able to map them it to the projects. A
setter is added for the URN of the project model so that the establishment can
be assigned.