Fetch bulk Establishments and Trusts on the Projects index page
This calls the EstablishmentsFetcher and IncomingTrustsFetcher services from
the project index action. The bulk establishments and trusts are assigned to
the projects, and as these are memoized, requests to fetch individual
establishments and trust will no longer be made.

The services are mocked in the feature specs. While this is not ideal, mocking 
the endpoints was adding a significant amount of noise and complexity to 
the tests.