Update factories for converison details
Our model is getting complicated and we need our factories to help us as
much as possible. This work refactors some of the factories.

We define two empty `Conversion::Details` factories and create two
`Conversion::Project` factories.

Right now, all projects are actually voluntary conversion projects, so
we user an alias to keep the suite green:

`conversion_project` == `voluntary_conversion_project`

Whislt doing this work we had issues with the two factories when
creating the associated Conversion::Details objects - we would end up
with two objects and specs randomly failing when the wrong details
object was returned.

By splitting these into two distinct factories we can stop this at the
cost of some duplication.

We have also made some changes that we hope align to the Factory best
practices, making a generic project factory as simple as possible and
adding traits as needed:

https://github.com/thoughtbot/factory_bot/blob/main/GETTING_STARTED.md#best-practices