Add base `TaskList::Task` model
We are adding a new task list implementation. This requires a base task
model that defines a key based on the class name.

The key will be to enable dynamic behaviour, such as:
- the task list route parameter.
- a reference to any associated task notes (this will not be as robust as a
  proper foreign key, developers will need to be cautious when renaming any
  tasks).
- a key to map to and from a backing task list ActiveRecord model that contains
  all of the task list attributes.

The task model also computes the status of the task. The `in_progress?`,
`completed?`, and `not_applicable?` methods can be overridden for any tasks
that are more complex.

This is namespaced under `TaskList`, and will be inherited by the tasks of each
separate task list.