Submit the Project create form to the FormObject
When a user fills in the project create form in the UI, the form will now submit
to the relevant Voluntary/Involuntary FormObject. This FormObject then handles
all the creation/tasklist assignment/note creation tasks.

The signature of `notify_team_leaders` has changed because we need to explicitly
pass it the created project, not the project form.