Add `TaskList::Base` class and basic `Conversion::Voluntary::TaskList`
We want an easy way to create different types of task list by defining a
task list layout composed of `TaskList::Task` objects.

The `TaskList::Base` class handles the mapping of the `TaskList::Task` objects
to the task list database table. This uses the `TaskList::Task` key/identifier,
which must be named consistently with the database columns. For example, a task
list containing a 'TaskList::Task` named
`Conversion::Voluntary::Tasks::Handover` which has a `review` attribute, must
have a column named `handover_review`.

For now, the base class is tested through an instance of a
`Conversion::Voluntary::TaskList`. As we introduce more task list types, this
will likely be moved into a shared spec.