Add support for Project > TaskList association
We'll be using delegated types [1] for this, a sort of 'reverse
polymorphic associtation'.

For us, this means:

A Project can have a task list and each task list can be a different
type. Right now we have a voluntary converison task list and a
involuntary converison task list.

Rails will use these two columns to manage the association.

The choice of 'task_list' goes against the recommended 'task_listable'
as we find task_list more approachable and cannot see a reason why we
wouldn't want the interface to looks like:

`project.task_list`

[1] https://edgeapi.rubyonrails.org/classes/ActiveRecord/DelegatedType.html