Add delegated type to Project
The previous commit setup the columns to store this association.

Here we tell rails what the association is know as and as mentioned,
decided to not follow the recommendation that would have been
'task_listable' and go with 'task_list'.

We like how this makes the interface nice, at the cost of making it less
obvious that this is a delegated type.