Add involuntary conversion task list
We've already proven the delegated type on Project is working for a
single task list type.

Here we add a second task list type, involuntary conversion, to give
us the confidence that this is all working as we intend.

Right now we are not doing anything with the
Converison::Involuntary::TaskList, that comes later.