Fix bug where empty URNs and UKPRNs raise an API error
Currently, when creating a project with an empty URN or UKPRN, the project
model will try to validate these against the API. The API will throw a bad
request error, which is re-raised in the API client, and the user will see an
error page.

Instead, we can conditionally validate that the URN or UKPRN returns an
establishment or trust only if they are present. This ensures that a
blank URN or UKPRN will not call out to the API for validation, and will
instead show the error message associated with `presence: true`.