Refactor readme and documentation
As we move into running a live service and because healthy documentation
is a good thing, we want to keep our readme focused.

Here we create more in-depth documentaiton and link from the readme, the
intention is to help layer the documentation from light in the top level
readme to very in-depth, for example in the Terrafrom specific readme.