Do not allow negative numbers in dates
While negative numbers are perfectly valid in Ruby dates (e.g.
`Date.new(-1, -1, -1)` parses to `-0001-12-31`) the underlying database
cannot handle them, and trying to submit a date with negative numbers results
in a `TinyTds::Error`.

Catch any negative numbers in dates and treat the date as invalid if any are
present.