Make it explicit where the console & log sections are
When trying to log into the Development site for the first time, it wasn't
obvious at first the the Console & Log access sections were below the PIM
section. It's easy to miss them if you're trying to access Development
and don't need a PIM.

Add a couple of links to the "Development" paragraph to show where the console
and log access sections are.