Project provisional conversion date can be in the past
We are loosening the validations on the Project model, now that we use a
FormObject to create projects, which handles the validations on creation.

The project provisional conversion date does not need to be enforced as being
in the future, as the FormObject handles this on creation.

However, removing the future validation uncovered a bug. There is an additional
validation which ensures the provisional conversion date the 1st of the month. If
"today" is not the first of the month, the existing test checking that the date
is not in the future throws an unexpected error saying that "today" is not the
first of the month.

As we no longer care if the provisional conversion date is today or not, we can
remove this whole block.