Add trust_exists & establishment_exists validations
Add the trust_exists & establishment_exists validations to the FormObject.

With help from https://github.com/DFE-Digital/dfe-complete-conversions-transfers-and-changes/pull/303/

This adds validations to the FormObject which check if the trust & school
actually exist, before creating the project. It also tightens up some of the
existing validations, and corrects a potential false positive in the existing
tests for `urn` and `ukprn`.