Task lists can return their locales_path
We want all the content for task list to be scoped to the project type, i.e.
voluntary or involuntary conversions so that we can have unique content
for each.

Our first step is to allow the task list to return a `locales_path` which
we'll use to strucutre the locales files.

Co-authored-by: Co-authored-by: Lorna <lorna@dxw.com>