Use locales to render task list index
Now that we have the bare bones of task list objects return locale
paths, we can utilise them in our views.

Here we update the shared task list index view to use the locales. This
view also makes use of a helper method to produce a html id for use in
the markup, in order to retain the existing task list and the new one we
had to introduce a second helper to do the same thing from the locales.

We will have to remove the old one along with the old task list when the
time is right.