Add `task_identifier` column to the notes table
We currently have notes that are associated with a `Task` model through a
foreign key.

Our new `TaskList::Task` model is not backed by a table, so we need another way
to associate notes with tasks.

This adds a `task_identifier` column to the notes table that will hold a
reference to a `TaskList::Task.identifier`.

We are aware that there is now a risk of orphaned notes when renaming tasks, or
when deleting tasks. We have agreed that this is preferable to managing an
ActiveRecord model for each task in the task list. Future developers will have
to be cautious when renaming or deleting, and we will provide guidance on this.