Use view components for common task elements
The task list is made up of lots of tasks with idividual 'actions' to
complete.

We want to make managing these task views simple. As we already use the
GOVUK View Components [1] it makes sense to create own own view
components for our own common UI.

Here we have a simple TaskHeaderComponent that renders the header
information and a CheckBoxActionComponent; Right now, almost all actions
will use the CheckBoxActionComponent.

We decided on this view component approach for these actions right now,
another approach would be to extend the GOVUK form builder [2] with our
form elements - this felt like something we could look to do later.

[1] https://govuk-components.netlify.app/
[2] https://govuk-form-builder.netlify.app/

Co-authored-by: Lozette <laura@dxw.com>