DRY up notes return destination
The back links on the notes pages, and the redirect on `Note` create and update
currently go to the same destinations (the project task page for project notes,
and the task page for task-level notes).

We want to introduce notes to the new task list (alongside the old task list),
which mean we will need to add a new destination. 

DRYing up this into a single helper will enable us to more easily make those
changes.