Remove CheckboxAction component dependency on form
We found when trying to test the Checkbox Action component that its dependency
on a form made testing extremely difficult. Instead of passing in a form to the
component, it makes it easier to test if we instad manually render the HTML
the form *would* have rendered.

This commit also includes some html and css additions to render the checkboxes
in the Govuk Formbuilder "style".