Ensure the task identifier is stored as NULL when absent
When no value is provided, Rails will store task_identifier as an empty string.
Instead, we want to ensure these are stored as NULL, so that we aren't mixing
blanks and NULLs.

There are libraries that enable this behaviour (such as `nilify_blanks`), but
it doesn't feel worth introducing a dependency yet.