Add nonce to inline scripts so that are not blocked by the CSP
We have added a CSP to the application. 

This is blocking our inline scripts for:
- Enabling the GOV.UK JavaScript
- Enabling the Accessible Autocomplete component.

This adds nonces to the inline scripts. Because a nonce generator is configured
in `content_security_policy.rb`, we only need to call
`content_security_policy_nonce`.

Note: It's nicer to add nonces to a `javascript_tag` (for example `<%=
javascript_tag nonce: true do %>`). However, the ERB linter suggests avoiding
`javascript_tag` altogether.