Add validation when both `not_applicable` and other actions are selected
Users should not be able to select both "not applicable" and any other
actions.

Add validation to the optional task model, derived models must define an error
key and message.

It's important to note that the attributes of an optional task must be cast
(via the attribute method on the task model) for the validation and subsequent
re-rendering of the page, else they get assigned as "0" and "1" (which are both
true), and we have to handle the cast manually.