Add missing Not applicable checkbox to tasks
While we were doing this work, the work to create Not applicable boxes in the
new tasks lists was completed. Go back and add Not applicable checkboxes to
tasks which are now missing them.

Change the base class for these tasks to TaskList::OptionalTask to indicate they
use a `not applicable` box.