Use `write_attribute` in Note setter
In some instances `super` here will not be what you think it is - this
seems to always be true in production like environments.

Before this change, calling `super(value)` results in:

```
NoMethodError (super: no superclass method `task_identifier=' for #<Note>
```

Calling this in local development does work and we have not been able to
identify why.

The fix here is to use the `write_attribute` method to achieve the same
end result without using `super` i.e to set the attribute on the model.

We can only validate this fix will work in our development environment.