Add style for a task section hint
We have the a single task view that often get broken down into
'sections'.

A section can have hint text that should be styled as such and was
previously missing.

Whilst I would like to rationalise all of the styles for the task page,
it will have to wait, this adds a style that we can use when the
content is wrapped in a `app-task-section__hint` div.