Add conversion projects scope
As we head towards a world with different project types, it makes sense
to have an interface to those types.

Here we add a conversions scope on Project that on includes conversions
projects, whilst this has little value right now as we have no other
types, it will form the foundation for the voluntary and involuntary
conversion projects scopes that will follow.