Refactor project views
We are heading towards a world where there are more than one type of
projects. Our existing project is really a conversion project and we
have two more types transfers and changes in the pipeline.

We want to be able to render index views of projects of different types,
here we refactor the way project index is composed so that it can be
reused across different views.