Redirect to 404 page if project cannot be found
As we introduce more types of projects there will be instances when a
project exists but cannot be found. This is because we scope the
projects down to a subset and then find the id.

For example, a voluntary project exists but should not be shown on an
involuntary path.

By adding `rescue_from` to the base `projects_controller` we get this
behaviour in all subclasses.