Move the new involuntary task list
This was at a temporary path, we want the task lists to live at:

`/conversions/involuntary/projects/id/task-list`

To keep the test suite green this is a pretty big commit, however the
majority of this is updating Rails path helpers.

We added a helper to return the path to a given involuntary conversion
project so that it only needs to be set in one location.