Switch to the new task list
By changing the links on the shared project list view, when a users
follows a link to a project, they are taken the appropriate project show
page.

Once there we also update the project sub-navigatio to link to the new
paths.

Finally we make sure the project summary and sub-navigatio is show on
all the new task list views.

With these changes the application has been switched to the new task
lists and access to the old one is only via the old paths.