Refactor task locale feature tests to be smaller
I did consider using shared examples for these, but it would make debugging
pages with missing locales a lot harder. Shared examples make it harder to see
which line the failure occurred on, and pinpoint which locale is missing strings.

In this iteration the missing locales will be indicated like so:
`has locales for Articles of association (FAILED - 1)`