Compare task count on task list in UI, to the array of tasks to check
When a new task is added to the task list, we will need the task name to be
added to the `tasks` array in this spec. To ensure this happens, we can change
the test in this spec which checks the number of tasks, so that is compares the
number of tasks on the page with the number of tasks in the test array. If
these fall out of sync, we know a new task has been added (or taken away) and
the specs need updating.