Add a not found error to the base controller
Lots of controllers might need to render a not found error so putting
this here gives them all access should they need it.

We remove the incorrect version of this from the projects_controller as
it can use this new one.

We also switched to 'inline private' definitions as this is consistent
across the code base.