Task list contollers not found error
When a single project is shown we redirect to the task list path,
essentially makeing this action the 'show project' action.

Whenever we show a single project, it should be from a scoped list of
projects, i.e voluntary projects only show up on voluntary paths and not
involuntary ones.

Here we resuce from ActiveRecord::RecordNotFound in task list
controllers and scope the voluntary and involuntary controllers
appropriately.