Update rediection paths after assigment
Once a users assigns another users to a project we want to redirect to
the appropriate path. Whilst we still support the generic project view
which can handle this, it is better to redirect to the correct path.

Here we switch to our project path helpers, using these in the
controller means prefixing with the `helpers` object - this is not ideal
but is probably good enough for now.