Redirect to correct path after create project
When a new project is created we want to take the user to that project
on the correct path, whilst the generic projects paths can handle this
right now, this may not always be the case.

I couldn't find a good way to test this with the shared example so I
settled on this approach - I am not loving it, but I think it will have
to do given time constriants right now.