Update contact category options
We had user feedback when adding a new contact:

- having 'other' as the first and default option was confusing
- that a 'solicitor' option was missing

This commit resolves this by adding a 'Choose category' option when
creating a new contact and adding the missing category.

We also removed the `format_category_name` helper as we can use Rails
humanize method.

We did not add the default 'Choose category' option on the edit view as
it felt unnecessary there as a value is pre-selected.