Validation for contact category
We can't accept the default value of 'Choose categoty' when adding a new
contact, here we add validation and errors to help users.

We cleanup the validation specs on the Contact model, we use shoulda
matchers and the `validates_presence_of` matcher checks for empty/nil
values by default, we don't need to test them seperately.