Control the order of contacts
Having learnt that the order of the contact groups is important to user
it follows we should keep the order consitent.

Here we take over control of the ordering when rendering a list of
contacts using a new scope.

This work also removes the unnecessary `where(project)` part of the
query.