Update contact redirect and back paths
We want to redirect to the appropriate page after adding a new contact.

In the spec we make it clear we are dealing with a voluntary conversion
(the converison_project factory is actually voluntary) and test the path
against that.

As with all the paths, it's getting complicated to mangage them and we
need to pay off that debt, but not as part of this work.