Rollback Sidekiq
We discovered that the version of Redis running in Azure is not yet
compatible with Sidekiq 7.

As we don't use any Sidekiq 7 features and there is a block on upgrading
Redis in Azure for at least 10 days [1], we think it best to rollback
and lock sidekiq to the 6.x branch for the moment.

[1] https://learn.microsoft.com/en-us/azure/azure-cache-for-redis/cache-how-to-upgrade#upgrade-using-the-azure-portal