Configure the rails session
Before this change the name of the cookie rails uses to store a users
session would be:
`_dfe_complete_conversions_transfers_and_changes_session`
which is cumbersome.

As our cookies are scoped by domain it makes sense
to have something shorter like SESSION.

As the azure cdn adds its own cookies which are all caps, it makes sense
to be consistent.
A shorter cookie name also makes the cookie page layout nicer.

After this change users will be forced to sign in.