Add the cookie banner
We render the cookie banner which also acts a form for users to accept
or reject optional cookies.
We update the `cookies_controller` to handle submissions from either form.
Once the cookie is set, we hide the banner.
When a user submits the cookie form we show a flash message to indicate
the action has been performed.
We used `redirect_back_or_to` to return the user to the page they
submitted the form from.

https://api.rubyonrails.org/classes/ActionController/Redirecting.html#method-i-redirect_back_or_to

If `redirect_back_or_to` cannot find the correct page, it falls back to
the cookie page.

We created a cookies concern because the banner is shown of every page
until the user makes a selection. This means we need to load
`@cookie_form` on every page.