Move conversion route to the Conversion::Project
Previously we had planned to store all involuntary or voluntary details
in the conversion_details table - this is no longer the case as
everything will be in the appropriate `task_lists` table.

Here we move the only attribute we do use in the details model,
'conversion route' into the conversion project model and once done we
can stop lazy loading details with projects.