Contacts are ordered as a user might expect
During a recent product review it was noted that the contact groups were
ordered in an unexpected way. We talked about this and decided that the
order users expect would likely be by the likelyhood of using the
contacts, so pretty arbitary.

Previously, we used `group_by`, an enumerable [1] method, not active
record in a scope. Revisiting this, it felt wrong for a scope to return a
hash, so we have changed the approach here. We also noted that group_by
does not honour sorting - which was resulting in a seemingly random
order of the contact groups noted.

Honestly, this felt far more difficult that it should be! We have relied
on the category order on the Contact model to define the expected order.
Some acceptably small logic in the view pulls out contacts from the
groups in the order defined in the class.

[1] https://ruby-doc.org/3.1.2/Enumerable.html#method-i-group_by