Test the Task generators by testing their output
Add specs to test the task generators. These work by running the generators,
and testing their output against expected examples in `fixtures`.

Unfortunately this does not satisfy the SimpleCov requirements, as SimpleCov
does not recognise that we're testing the methods in the generators.

However having done some searching, this appears to be the most accepted way
to test rails generators.