Do not check the Task generators with SimpleCov
We are testing the output of the Task generators in
`conversion_involuntary_task_generator_spec.rb` and
`conversion_voluntary_task_generator_spec.rb`

However SimpleCov does not recognise these specs as testing the generators,
because they do not directly call `copy_files`, `gsub_files` and
`generate_migration`.

So, we have to exclude these generators from SimpleCov, to keep the coverage
at 100%; even though we are testing the generators indirectly.