Improve the cookie preference flash message
According to the govuk design system, once a user has set their cookie
preferences we show a "hide banner".
We believe the purpose of the hide banner is to reminder users that they
can change their cookie preferences at anytime.

To get the behaviour we wanted we use the Rails flash message, having
the flash message felt like it was doing the job of the hide banner and
so was duplicating the bahaviour.

We talked this over as a pair and decided that the flash message was
enough as long as it guided the user.

Previously, all flash messages were rendered as a heading. We didn't
feel this was right for our message and changed it. We then reviewed all
flash messages and felt this change was appropriate.

We may want to review the design of flash messages at a later time.

https://design-system.service.gov.uk/components/cookie-banner/