Add notification banner component
We make use of the Rails flashes. Our general approach is to show one
after a successful operation such as:

- user adds a note
- user adds a contact
- user adds a project
- user signed out

We chose to use the Notfication banner component for this [1], did not
use the explicit 'success' version. This version is designed
specifically for

> Reacting to something the user has done

which is a much better fit for our use case.

We do have a couple of instances where the flash is used for
unsuccessful events, such as:

- user is not registered
- user could not be signed in
- user is not authorised

Here we introduce a view component to handle both cases, using a view
component helps us encapsulate all the behaviour and make it easy to
test.

The component takes the two standard Rails flash types notice and alert
and renders the appropriate design.

We also add styles to control the content of the flashes so that
developers do not have to.

[1] https://design-system.service.gov.uk/components/notification-banner/