Add support for Google Analytics
We were asked to add google analytics to the application to help us
understand how users interact with our application.

We had already laid out the ground work for accepting optional cookies,
which google analytics is one of.

Because analytics relies on Javascript loaded from a third party domain
we had to modify the content security policy. We weren't sure how to do
this, but followed google documentation here:

https://developers.google.com/tag-platform/tag-manager/web/csp

The domains we had to allow within the content security policy for
analytics didn't seem to be documented, so we used trial and error for
this process.

As all the allow domains are clearly google domains, we felt this was an
acceptable risk.

We only want to include analytics when:
- In production
- The google analytics id is present
- The user has accepted optional cookies

The enable_google_tag_manager method supports this.

We also update the cookie page to include information about the Google
analytic cookies.